factorize bench unstruct + remove compilation warning
